+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/metric/MetricLoader.java
@@ -2,6 +2,7 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.data.config.metric;
 
+import com.yahoo.bard.webservice.data.dimension.DimensionDictionary;
 import com.yahoo.bard.webservice.data.metric.MetricDictionary;
 
 /**
@@ -13,7 +14,19 @@ public interface MetricLoader {
      * Load metrics and populate the metric dictionary.
      *
      * @param metricDictionary  The dictionary that will be loaded with metrics
+     *
+     * @deprecated in favor of loadMetricDictionary(MetricDictionary, DimensionDictionary)
      */
+    @Deprecated
     void loadMetricDictionary(MetricDictionary metricDictionary);
 
+    /**
+     * Load metrics and populate the metric dictionary with dimension dictionary for dimension dependent metrics.
+     *
+     * @param metricDictionary  The dictionary that will be loaded with metrics
+     * @param dimensionDictionary  The dimension dictionary containing loaded dimensions
+     */
+    default void loadMetricDictionary(MetricDictionary metricDictionary, DimensionDictionary dimensionDictionary) {
+        loadMetricDictionary(metricDictionary);
+    }
 }